update mock-doc performance function signatures
update the function signatures of mock-doc's performance implementation
by declaring the return type on `mark()` and `measure()` to align with
the `Performance` interface, but _continue to return undefined_,
ignorning the type checker errors.

stencil's stubs for these functions assume access to underlying browser
api's (namely a performance buffer). given that older browsers may not
support this functionality and implementating a mock version of these
functions falls well outside the scope of a typescript update, simply
ignore the errors.

we take this approach, rather than:
```typescript
// @ts-ignore
mark(): void { /* no-op */ }
```
as stencil can (and does) instantiate an instance of `MockPerformance`
and attempts to assign it to variable(s) that are of type
`Performance`. this circumvents the type checker (although stencil was
lying to it already) to prevent additional `// @ts-ignore` pragmas

this may result in typing errors for consumers, however the actual
values shall remain the same.